// backend/types/MessageAction.ts
export enum MessageAction {
  GET_STATE = "getState",
  ADD_USER = "addUser",
  UPDATE_USER = "updateUser",
  GET_USER = "getUser",
  ADD_ADDRESS_USER = "addAddress",
  UPDATE_ADDRESS_USER = "updateAddress",
  GET_ADDRESSES_USER = "getAddresses",
  LIST_PROFILES = "listProfiles",
  LIST_PERMISSIONS = "listPermissions",
  MANAGE_PROFILE_PERMISSIONS = "manageProfilePermissions",
  LIST_PROFILE_PERMISSIONS = "listProfilePermissions",
  ADD_ACCOUNT = "addAccount",
  UPDATE_ACCOUNT = "updateAccount",
  DELETE_ACCOUNT = "deleteAccount",
  LIST_ACCOUNTS = "listAccounts",
  COUNT_PAGES_ACCOUNTS = "countPagesAccounts",
  MANAGE_ACCOUNT_LINKED_USERS = "manageAccountLinkedUsers",
  LIST_ACCOUNT_LINKED_USERS = "listAccountLinkedUsers",
  COUNT_PAGES_USERS = "countPagesUsers",
}